refactor(agent): split ViewRenderer into PromptRenderer and DashboardRenderer
Split the 827-line View.Renderer into focused modules:

- View.PromptRenderer: prompt input box rendering, cursor positioning,
  and layout helpers (used by AgentMouse for hit-testing)
- View.DashboardRenderer: dashboard sidebar (context, model, LSP, directory)
- View.RenderInput: shared data extraction struct (fixes nested module violation)

Updated all callers (content.ex, agent_mouse.ex) and split tests accordingly.

Closes #642
